visualizer/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
python-dotenv==0.15.0
requests==2.25.1
psycopg2-binary==2.8.6
praw==7.2.0
celery
django-celery-beat
redis
pytest-django==4.1.0
factory_boy==3.2.0
